restaurant-system/requirements.txt
flask
flask-socketio
python-socketio
eventlet
tk
requests
websocket-client